+++ src/briefy/ws/resources/events.py
@@ -1,32 +1,25 @@
 """briefy.ws.resources events."""
+from briefy.common.db.model import Base
 from briefy.common.event import BaseEvent
 from briefy.common.event import IDataEvent
 from briefy.ws import logger
+from pyramid.request import Request
 from zope.interface import implementer
 
 
 class BaseResourceObjectEvent:
     """Base class for object events: load and delete."""
 
-    def __init__(self, obj, request, **kwargs):
+    def __init__(self, obj: Base, request: Request, **kwargs):
         """Custom init to call parent BaseEvent if it exist."""
         self.request = request
         self.obj = obj
-        try:
-            super().__init__(obj, **kwargs)
-        except TypeError:
-            pass  # TODO: fix this to look to __mro__ and use different call
 
-    def __call__(self):
+    def __call__(self) -> str:
         """Notify about the event, need to be implemented by subclass.
 
-        :returns: Id from message in the queue
-        :rtype: str
+        :returns: A string
         """
-        try:
-            super().__call__()
-        except AttributeError:
-            pass  # TODO: fix this to look to __mro__ and use different call
         return ''
 
 
@@ -48,14 +41,14 @@ class ResourceObjectEvent(BaseResourceObjectEvent, BaseEvent):
 
     logger = logger
 
-    def __init__(self, obj, request):
+    def __init__(self, obj: Base, request: Request):
         """Custom init to call parent BaseEvent if it exist."""
         user = getattr(request, 'user', None)
+        user_id = None
         if user:
             user_id = user.id
-        else:
-            user_id = None
-        kwargs = dict(actor=user_id, request_id=None)
+
+        kwargs = {'actor': user_id, 'request_id': None}
         super().__init__(obj, request, **kwargs)