+++ HISTORY.rst
@@ -6,6 +6,8 @@ History
 ------------------
 
     * Improving query and count to be cached in the view class to avoid query duplication in case of a second call (rudaporto).
+    * Upgrade depencies to new versions and in special pyramid to 1.9.1 (rudaporto).
+    * Improving query and count to be cached in the view class to avoid query duplication in case of a second call (rudaporto).
     * Increase usage of type hints (ericof).
     * Increase code coverage (ericof).