+++ HISTORY.rst
@@ -10,6 +10,7 @@ History
     * Improving query and count to be cached in the view class to avoid query duplication in case of a second call (rudaporto).
     * Increase usage of type hints (ericof).
     * Increase code coverage (ericof).
+    * Fix Workflow imports (ericof).
 
 2.0.6 (2017-07-27)
 ------------------