+++ HISTORY.rst
@@ -11,6 +11,11 @@ History
     * Increase usage of type hints (ericof).
     * Increase code coverage (ericof).
     * Fix Workflow imports (ericof).
+    * New base resource to return a paginate collection from a custom plain sql query (rudaporto).
+    * Refactor all mocks used for resource tests (rudaporto).
+    * Added new test to check RESTService.collection_get method (rudaporto).
+    * Move database fixtures to conftest and use Base sqlalchemy from briefy.common (rudaporto).
+
 
 2.0.6 (2017-07-27)
 ------------------