+++ HISTORY.rst
@@ -20,6 +20,7 @@ History
     * Validate if we can find the column to do the filer and send the proper error messsage if not (rudaporto).
     * Make sure we call super().__init__ and super().__call__ always and logger as debug if some error happens (rudaporto).
     * Disable by default the 500 server error view in test and development environments (rudaporto).
+    * Handle ValidationError on WorkflowAwareResource.post (rudaporto).
 
 
 2.0.6 (2017-07-27)